Only allow creation of original 150 pokemon
This PR ensures that when a User creates a pokemon by typing in their
name, they can only create a pokemon that was part of the original
150.

In order to achieve this the codebase needs to know about the
original 150. I have encapsulated this data into a PORO, that I will
then be able to use in the future for other validations such as making
sure when a pokemon evolves it will evolve into the correct pokemon.